requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
greenlet

